+++ src/svd/access.rs
@@ -1,10 +1,8 @@
 use xmltree::Element;
 
 use crate::elementext::ElementExt;
-#[cfg(feature = "unproven")]
 use crate::encode::Encode;
 use crate::error::*;
-#[cfg(feature = "unproven")]
 use crate::new_element;
 use crate::types::Parse;
 
@@ -36,7 +34,6 @@ impl Parse for Access {
     }
 }
 
-#[cfg(feature = "unproven")]
 impl Encode for Access {
     type Error = anyhow::Error;
 
@@ -54,7 +51,6 @@ impl Encode for Access {
 }
 
 #[cfg(test)]
-#[cfg(feature = "unproven")]
 mod tests {
     use super::*;
     use crate::run_test;