+++ src/svd/register.rs
@@ -4,9 +4,8 @@ use xmltree::Element;
 
 use crate::types::Parse;
 
-#[cfg(feature = "unproven")]
 use crate::elementext::ElementExt;
-#[cfg(feature = "unproven")]
+
 use crate::encode::Encode;
 use crate::error::*;
 use crate::svd::{dimelement::DimElement, registerinfo::RegisterInfo};
@@ -52,7 +51,6 @@ impl Parse for Register {
     }
 }
 
-#[cfg(feature = "unproven")]
 impl Encode for Register {
     type Error = anyhow::Error;
 
@@ -70,7 +68,6 @@ impl Encode for Register {
 }
 
 #[cfg(test)]
-#[cfg(feature = "unproven")]
 mod tests {
     use super::*;
     use crate::dimelement::DimElementBuilder;