+++ src/svd/registerinfo.rs
@@ -1,13 +1,11 @@
-#[cfg(feature = "unproven")]
 use std::collections::HashMap;
 
 use crate::elementext::ElementExt;
 use xmltree::Element;
 
-#[cfg(feature = "unproven")]
 use crate::encode::Encode;
 use crate::error::*;
-#[cfg(feature = "unproven")]
+
 use crate::new_element;
 use crate::parse;
 use crate::types::Parse;
@@ -244,7 +242,6 @@ impl RegisterInfo {
     }
 }
 
-#[cfg(feature = "unproven")]
 impl Encode for RegisterInfo {
     type Error = anyhow::Error;
 
@@ -334,7 +331,6 @@ impl Encode for RegisterInfo {
 }
 
 #[cfg(test)]
-#[cfg(feature = "unproven")]
 mod tests {
     use super::*;
     use crate::run_test;