+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix: panic with 32 bit wide fields that have enumerated values
 - Fix: produce error on 0-width fields
 - Fix: error instead of panic when an array/cluster name is missing the `%s` placeholder
+- [breaking-change] Add support for 64 addresses, register values, enum values and writeConstraints
 
 ## [v0.9.0] - 2019-11-17