+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.10.0] - 2021-04-03
+
+- Added `strict` feature that hides part of checks
 - Builder pattern now is used for creating structures
 - Peripherals are processed in parallel with `rayon`
 - Serde now skips empty tags
@@ -126,7 +129,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Initial SVD parser
 - A `parse` utility function to parse the contents of a SVD file (XML)
 
-[Unreleased]: https://github.com/rust-embedded/svd/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/rust-embedded/svd/compare/v0.10.0...HEAD
+[v0.10.0]: https://github.com/rust-embedded/svd/compare/v0.9.0...v0.10.0
 [v0.9.0]: https://github.com/rust-embedded/svd/compare/v0.8.1...v0.9.0
 [v0.8.1]: https://github.com/rust-embedded/svd/compare/v0.8.0...v0.8.1
 [v0.8.0]: https://github.com/rust-embedded/svd/compare/v0.7.0...v0.8.0